Fixed compile error from wrong data types, only seems to be a problem in NUView on OS X.

+++ Vision/Modules/goaldetection.cpp
@@ -377,7 +377,7 @@ void GoalDetection::detectGoal(ClassIndex::Colour colour, vector<Quad>* candidat
             counter = 0;
             for (unsigned int j = 0; j < start_trans.size(); j++)
                 if (start_trans.at(j).getLocation().x >= start_pos && start_trans.at(j).getLocation().x <= end_pos) {
-                    sdev += pow(start_trans.at(j).getLocation().x - mean, 2);
+                    sdev += pow(static_cast<float>(start_trans.at(j).getLocation().x - mean), 2);
                     counter++;
                 }
             sdev = sqrt(sdev/counter);
@@ -408,7 +408,7 @@ void GoalDetection::detectGoal(ClassIndex::Colour colour, vector<Quad>* candidat
             counter = 0;
             for (unsigned int j = 0; j < start_trans.size(); j++)
                 if (end_trans.at(j).getLocation().x >= start_pos && end_trans.at(j).getLocation().x <= end_pos) {
-                    sdev += pow(end_trans.at(j).getLocation().x - mean, 2);
+                    sdev += pow(static_cast<float>(end_trans.at(j).getLocation().x - mean), 2);
                     counter++;
                 }
             sdev = sqrt(sdev/counter);
